Rename DB Output to Stylebook Output and simplify panel
Made-with: Cursor
