+++ apps/agate-ui/src/nodes/place_extract/PanelComponent.tsx
@@ -84,7 +84,7 @@ const nodeMetadata = {
   ]
 };
 
-import { getNodeOutputById } from '@/lib/nodeOutputs'
+import { getNodeOutputById } from '../../../node_outputs_ui'
 import { Label } from '@/components/ui/label'
 import { Textarea } from '@/components/ui/textarea'
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select'
