+++ packages/backfield-core/src/backfield_core/nodes/db_output/metadata.json
@@ -1,12 +1,12 @@
 {
   "type": "DBOutput",
-  "label": "DB Output",
+  "label": "Stylebook Output",
   "icon": "Database",
   "color": "bg-slate-500",
-  "description": "Persist consolidated upstream JSON into Backfield Postgres (worker-local)",
+  "description": "Persists results to Stylebook",
   "category": "output",
   "requiredUpstreamNodes": [],
-  "dependencyHelperText": "Same consolidation rules as JSON Output: wire any upstream nodes you want merged (unwraps node-* namespaces, apply include/exclude), then persists when `places` is present. JSON Output is optional — not required upstream.",
+  "dependencyHelperText": "Persists results to Stylebook",
   "inputs": [
     {
       "id": "data",
@@ -32,8 +32,5 @@
       "type": "string"
     }
   ],
-  "defaultParams": {
-    "exclude": null,
-    "include": null
-  }
+  "defaultParams": {}
 }
